# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGASTYA AI Dead Reckoning API",
    "description": "Asynchronous Neural-Inertial Navigation, ES-EKF Sensor Fusion, and Telemetry Engine.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/navigation/mode": {
      "post": {
        "tags": [
          "Navigation"
        ],
        "summary": "Set Navigation Mode",
        "description": "Set operating navigation mode.",
        "operationId": "set_navigation_mode_api_navigation_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/navigation/inject-fault": {
      "post": {
        "tags": [
          "Navigation"
        ],
        "summary": "Inject Fault",
        "description": "Inject runtime sensor fault.",
        "operationId": "inject_fault_api_navigation_inject_fault_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaultInjectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/simulation/scenario/{name}": {
      "post": {
        "tags": [
          "Simulation"
        ],
        "summary": "Select Scenario",
        "description": "Load and execute scenario.",
        "operationId": "select_scenario_api_simulation_scenario__name__post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/simulation/speed": {
      "post": {
        "tags": [
          "Simulation"
        ],
        "summary": "Set Speed",
        "description": "Set simulation playback speed multiplier.",
        "operationId": "set_speed_api_simulation_speed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationSpeedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FaultInjectionRequest": {
        "properties": {
          "fault_type": {
            "type": "string",
            "title": "Fault Type",
            "description": "gps_jamming, accel_bias_jump, gyro_bias_jump, vo_dropout"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "default": 1.0
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec",
            "default": 10.0
          }
        },
        "type": "object",
        "required": [
          "fault_type"
        ],
        "title": "FaultInjectionRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ModeChangeRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Operating mode: ai_enhanced_ekf, classical_ekf, pure_dr, ai_only"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "ModeChangeRequest"
      },
...
      "SimulationSpeedRequest": {
        "properties": {
          "speed_multiplier": {
            "type": "number",
            "maximum": 20.0,
            "minimum": 0.1,
            "title": "Speed Multiplier",
            "default": 1.0
          }
        },
        "type": "object",
        "title": "SimulationSpeedRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}